rathaxes: add the transmission code for very simple packets in the e1000 sample
Very simple packets are those which don't need to be fragmented nor checksummed.

The transmission code is defined in different chunks/methods of the TxRing type.
An e1000::xmit method template has been added to call the different methods on
the TxRing object in the e1000::Context object. The interrupt handler has been
reworked to handle transmission interrupts and cleanup the TxRing accordingly.

In parallel to the transmission code, the Socket::SKBuff type is now a real
Rathaxes type with its own debug and DMA operations methods.

Also, sorry, this changeset is mixed with other changes:

- Small improvements everywhere now that the compiler uses CNorm unstrict;
- Add the Device::AbstractDevice type; this Rathaxes represents the generic type
  used by the kernel to represent a device;
- Add the Socket::AbstractSKBuff type too to represent the SKBuff type used by
  the kernel;
- Likewise the Ethernet::AbstractDevice type represents the type used by the
  kernel for Ethernet devices;
- Socket::SKBuff is now a real Rathaxes type which aggregates a KernelSKBuff
  object.