rathaxes: change the abstract type notation in the e1000 sample
Starting with ra24db32bf134 Rathaxes types are generated differently: you
don't define entire structures anymore, but only the fields. And the
compiler generate a typedef'ed structure on top of it.

This works fine, except for "abstract types" (the types defined by ?and
used to interact with? the kernel). We need to define these types as
Rathaxes types to interact with them, in Rathaxes, but with
ra24db32bf134 it means that we loose (hide) the original type from the
kernel, making it very difficult to use "abstract types" with the
kernels APIs.

For example, here the hypothetical generated struct for the abstract
type "struct net_device" from Linux would be:

typedef struct {
    struct net_device   data;
} rtx_GeneratedType;

And here is how we have to use it:

rtx_GeneratedType *my_struct;
kernel_api_function(&my_struct->data);

This &my_struct->data is actually always a nop, but it's confusing and
hard to understand.

This changeset changes the notation to:

rtx_GeneratedType *my_struct;
kernel_api_function((struct net_device *)my_struct);

Which is, I believe, more intuitive and coherent with how you would
initialize the my_struct pointer here with a return value from the
kernel (i.e: with a cast into rtx_GeneratedType *).